Added Rhodes, etc
Added Rhodes, since while it was aligned with Ptolemy it was not
controlled by him, or any of the other Diadochi. I've also made it so
that Achaemenids/Macedon have fluid conquests on each other, and added
the brief period in which Thrace was in control of the Kingdom of
Macedon.